+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 5.1.1
+- Fixed saved page lagging for huge lists
+- Fixed inconsistent ui in onboarding
+
 ## 5.1.0
 - Added new Audio sensitive lyrics backgrounds Gradient and Wave (#182)
 - Added pulsating buttons on audio sensitive backgrounds